mejorando el tema del cron

+++ routes/web.php
@@ -36,12 +36,14 @@ Route::get('cron', function(){
 
     foreach ($customers as $data) {
         $server = $data->server;
-        $plex->setServerCredentials($server->url, $server->token);
-        if(isset($data->invited_id) and !empty($data->invited_id)){
-            if(strtotime($data->date_to) < strtotime(date('Y-m-d'))){
-               $plex->provider->removeFriend($data->invited_id);
-               DB::table('customers')->where('id',$data->id)->update(['status'=>'inactive']);
-               $total++; 
+        if(!empty($data->server->url) && !empty($data->server->token)){
+            $plex->setServerCredentials($server->url, $server->token);
+            if(isset($data->invited_id) and !empty($data->invited_id)){
+                if(strtotime($data->date_to) < strtotime(date('Y-m-d'))){
+                   $plex->provider->removeFriend($data->invited_id);
+                   DB::table('customers')->where('id',$data->id)->update(['status'=>'inactive']);
+                   $total++; 
+                }
             }
         }
     }
@@ -53,12 +55,14 @@ Route::get('cron', function(){
     $demos = Demo::all();
     foreach($demos as $demo){
         $server = $demo->server;
-        $plex->setServerCredentials($server->url, $server->token);
-        if(isset($demo->invited_id) and !empty($demo->invited_id)){
-            if(strtotime($demo->end_date) < strtotime(date('Y-m-d H:i:s'))){
-               $plex->provider->removeFriend($demo->invited_id);
-               DB::table('demos')->where('id',$demo->id)->delete();
-               $total_demos++; 
+        if(!empty($data->server->url) && !empty($data->server->token)){
+            $plex->setServerCredentials($server->url, $server->token);
+            if(isset($demo->invited_id) and !empty($demo->invited_id)){
+                if(strtotime($demo->end_date) < strtotime(date('Y-m-d H:i:s'))){
+                   $plex->provider->removeFriend($demo->invited_id);
+                   DB::table('demos')->where('id',$demo->id)->delete();
+                   $total_demos++; 
+                }
             }
         }
     }
